- 소켓 이벤트 useChatSocket으로 옮김 - CHATUSER_STATUS 이벤트를 받으면 roomId의 해당 유저들 리스트만 업데이트 했으나 recoil로 접근시 roomId가 0이 찍혀(이유모름) 그냥 chat을 전부 인벨리데이트함 -> 비효율적임 수정 필요
