+++ frontend/app/src/view/ChatPanel.tsx
@@ -63,19 +63,7 @@ export const ChatPanel = ({ chats, leaveRoom }: PanelProps) => {
     'list',
   ])
   const mydata = chatusers?.find((user) => user.user.uid === me?.uid)
-  useEffect(() => {
-    if (socket === undefined) {
-      return
-    }
-    socket.on('CHATUSER_STATUS', (res) => {
-      console.log(res)
-      queryClient.invalidateQueries(['user', 'me'])
-      queryClient.invalidateQueries(['chat', roomId, 'list'])
-    })
-    return () => {
-      socket.removeAllListeners('CHATUSER_STATUS')
-    }
-  }, [socket])
+
   return (
     <Grid container justifyContent="space-between">
       <Grid item xs={8}>
