+++ frontend/app/src/view/ChatView.tsx
@@ -42,7 +42,6 @@ export const ChatView = ({ messages, setMessages }: Props) => {
   const [modal, setModal] = useState(false)
   const [_, setSelectedChat] = useRecoilState(selectedChatState)
 
-  const { data: me, isSuccess } = useUserQuery(['user', 'me'])
   const { data: chatRoomList } = useApiQuery<Room[]>(['chat', 'joinlist'])
   const { data: joinedRoomList } = useApiQuery<JoinedRoom[]>(['chat', 'me'])
 
@@ -51,64 +50,7 @@ export const ChatView = ({ messages, setMessages }: Props) => {
     setSelectedChat((selectedChat) => ({ ...selectedChat, bool: false }))
   }
 
-  useEffect(() => {
-    if (!isSuccess || socket === undefined) {
-      return
-    }
-    const { uid } = me
-    socket.on('NOTICE', (res: Message) => {
-      console.log(res)
-      if (res.senderUid === uid) {
-        queryClient.invalidateQueries(['chat', 'me'])
-        if (res.msgContent === 'banned')
-          setSelectedChat((selectedChat) => ({ ...selectedChat, bool: false }))
-      }
-      queryClient.invalidateQueries(['chat'])
-    })
-    return () => {
-      socket.removeAllListeners('NOTICE')
-    }
-  }, [me, socket])
-
-  useEffect(() => {
-    if (socket === undefined) {
-      return
-    }
-    socket.on('RECEIVE', (res: Message) => {
-      const id = res.roomId
-      const msg = {
-        ...res,
-        createdAt: new Date(res.createdAt),
-      }
-      console.log('incoming message')
-      console.debug(msg)
-      setMessages((messages: Messages) => {
-        return {
-          ...messages,
-          [id]: messages[id] ? [...messages[id], msg] : [msg],
-        }
-      })
-    })
-    return () => {
-      socket.removeAllListeners('RECEIVE')
-    }
-  }, [socket])
-
-  useEffect(() => {
-    if (socket === undefined) {
-      return
-    }
-    socket.on('DESTROYED', () => {
-      queryClient.invalidateQueries(['chat', 'me'])
-      queryClient.invalidateQueries(['chat', 'joinlist'])
-      setSelectedChat((prev) => ({ ...prev, bool: false }))
-    })
-    return () => {
-      socket.removeAllListeners('DESTROYED')
-    }
-  }, [socket])
-
-  if (!isSuccess || socket === undefined) {
+  if (socket === undefined) {
     return null
   }
